remove excess line breaks, add argument name for @Parameterized.Parameters, variable testData replace to beginning of the class, packages rename due to the naming rules
