+++ src/main/java/ru/yandex/practicum/scooter/pages/OrderPage.java
@@ -1,67 +1,48 @@
-package ru.yandex.practicum;
+package ru.yandex.practicum.scooter.pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-
-
 import java.util.List;
 import java.time.LocalDate;
 import java.util.concurrent.TimeUnit;
 
 public class OrderPage {
-
     private WebDriver driver;
 
     // Форма 1 - Для кого самокат
-
     // Поле ввода Имя
     private By nameInput = By.xpath(".//input[@placeholder='* Имя']");
-
     // Поле ввода Фамилия
     private By surnameInput = By.xpath(".//input[@placeholder='* Фамилия']");
-
     // Поле ввода Адрес
     private By addressInput = By.xpath(".//input[@placeholder='* Адрес: куда привезти заказ']");
-
     // Список Станция метро
     private By metroStationInput = By.className("select-search__input");
-
     // Элементы списка Станция метро
     private By metroStationList = By.xpath(".//li[@class='select-search__row']/button");
-
     // Поле ввода Телефон
     private By phoneInput = By.xpath(".//input[@placeholder='* Телефон: на него позвонит курьер']");
-
     // Кнопка Далее
     private By nextButton = By.cssSelector(".Button_Button__ra12g.Button_Middle__1CSJM");
 
     // Форма 2 - Про аренду
-
     // Поле ввода Когда привезти самокат
     private By deliveryDateInput = By.xpath(".//input[@placeholder='* Когда привезти самокат']");
-
     // Кнопка раскрытия списка Срок аренды
     private By rentTimeInput = By.className("Dropdown-arrow");
-
     // Элементы списка Срок аренды
     private By rentTimeList = By.className("Dropdown-option");
-
     // Чек-боксы выбора цвета самоката
     private By colorCheckBoxes = By.className("Checkbox_Label__3wxSf");
-
     // Поле ввода Комментарий для курьера
     private By commentInput = By.xpath(".//input[@placeholder='Комментарий для курьера']");
-
     // Кнопка Заказать
     private By orderButton = By.xpath(".//div[@class='Order_Buttons__1xGrp']/button[text()='Заказать']");
-
     // Кнопка Да во всплывающем окне подтверждения заказа
     private By yesButton = By.xpath(".//button[text()='Да']");
-
     // Надпись Заказ оформлен во всплывающем окне оформленного заказа
     private By titleOrderCreated = By.xpath(".//div[text()='Заказ оформлен']");
-
 
     // Конструктор класса
     public OrderPage(WebDriver driver) {
@@ -123,10 +104,14 @@ public class OrderPage {
     }
 
     // Метод заполнения поля Комментарий
-    public void setComment(String comment) {driver.findElement(commentInput).sendKeys(comment);}
+    public void setComment(String comment) {
+        driver.findElement(commentInput).sendKeys(comment);
+    }
 
     // Метод клика по кнопке Заказать
-    public void clickOrderButton() {driver.findElement(orderButton).click();}
+    public void clickOrderButton() {
+        driver.findElement(orderButton).click();
+    }
 
     // Метод клика по кнопке Да во всплывающем окне подтверждения заказа
     public void clickYesButton() {
